# Initial work towards dynamically resolving enum variants.