--- a/src/adapter/optimizations/mod.rs
+++ b/src/adapter/optimizations/mod.rs
@@ -1,3 +1,4 @@
 pub(super) mod impl_lookup;
 pub(super) mod item_lookup;
 pub(super) mod method_lookup;
+pub(super) mod variant_lookup;