@CLIDragon 2025-07-28
This is ready for review. I'm not happy with the code duplication in `build_variant_name_index` but I can't see a way around it.