@CLIDragon 2025-08-07
After some review, resolving variants does appear to have the intended effect - it's simply smaller in magnitude than I was hoping for.