@obi1kenobi 2025-08-09
Yes, I'm seeing a 10x speedup on that lint with this branch patched in than without it.

```
unpatched:
        PASS [   0.885s]       major        enum_unit_variant_changed_kind
        PASS [   0.858s]       major        enum_unit_variant_changed_kind        
        PASS [   0.848s]       major        enum_unit_variant_changed_kind

patched w/ this branch:
        PASS [   0.082s]       major        enum_unit_variant_changed_kind
        PASS [   0.074s]       major        enum_unit_variant_changed_kind
        PASS [   0.083s]       major        enum_unit_variant_changed_kind
```