@CLIDragon 2025-08-09
I believe the improved queries all trigger the dynamic path, e.g.`enum_variant_missing` 

```graphql
CrateDiff {
    baseline {
        item {
            ... on Enum {
                visibility_limit @filter(op: "=", value: ["$public"]) @output

                importable_path {
                    path @output @tag
                    public_api @filter(op: "=", value: ["$true"])
                }

                variant {
                    kind: __typename @output

                    variant_name: name @output @tag
                    public_api_eligible @filter(op: "=", value: ["$true"])

                    span_: span @optional {
                        filename @output
                        begin_line @output
                        end_line @output
                    }
                }
            }
        }
    }
    current {
        item {
            ... on Enum {
                visibility_limit @filter(op: "=", value: ["$public"])
                enum_name: name @output

                importable_path {
                    path @filter(op: "=", value: ["%path"])
                    public_api @filter(op: "=", value: ["$true"])
                }

                variant @fold @transform(op: "count") @filter(op: "=", value: ["$zero"]) {
                    name @filter(op: "=", value: ["%variant_name"])
                }
            }
        }
    }
}
```